requirements.txt
fastapi
uvicorn
pydantic
langchain
langgraph
langchain-openai
dotenv
PyPDF2
python-docx
PyMuPDF
python-multipart